feat(lexer): add initial lexer steps
This commit adds the initial strucutre of the lexer.

- The readChar method give us the next character and advance the
position of the input as well as updates the ch to be read.
- The NextToken method looks at the current character in the channel and
returns a token based on which character we're examining. Before doing
so we call readChar to advance the position of the pointers.

This makes the initial tests pass.