feat(lexer): add string checker and default case to parse undeclared tokens
This commit adds the following capabilities:

- Identify undeclared tokens: the lexer can use the default branch of
the switch case to deal with the tokens that aren't recognized in the
declared cases, as well as throw something as ILLEGAL if we truly can't
handle the chars.

- Check for strings: the isLetter helper function checks if the param is
a letter, this fn has a large impact on the lang because it determines
what kind of chars can be used in identifiers and keywords.

- Handle indentifiers: the readIdentifier method reads the identifier
and advance the positions until it has to handle a non-leter-char, we
use it to set the Literal field of the current token. With the added
LookupIdent from the token package, we can check the keywords table to
see if the identifier is a keyword, if it isn't it sets its type to a
IDENT, meaning it's an user-defined identifier