feat(parser): add initial parser using a "Prat parser" strategy
This commits adds the initial strucutre of our top down operator
precedence parser - or just Pratt parser. Our monkey lang statements
basically follow the let <identifier> = <expression>; form.

The parser has three fields:

- l: a pointer to an instance of the lexer.
- curToken: similar to the position pointer the lexer has but instead of
pointing to a char it points to a token.
- peekToken: similar to the readPosition pointer, but it points to the
next token.

Following the Pratt strategy we build the root node of the AST first and
then we recursively iterate over the child nodes building them, by
calling other functions that know which AST node to construct based on
the current token.

We can have a glimpse of this implementation with the following funcs:

- parseLetStatement: it first constructs an *ast.LetStatement node with
the token it's currently sitting on (a token.LET token), and then it
advances the tokens while asserting the next tokens through expectPeek()
- expectPeek: it enforces the correctnes of the order of tokens by
checking the type of the next token, it only advances the tokens if the
type is correct.